Update errors and code order
Errors needed some updating and I also moved the hash pointer to be set as early as possible #weekendthoughts.  If there is an error then the hash pointer in the error should point to the message it's responding to.  Also simplified code now that I know go a lot better.