Allow consumers of datachannel to wait for ack on a stream data message
+ Creates a new function SendStreamDataMessageAndWaitForAck. It behaves
  exactly like SendStreamDataMessage but instead of returning
  immediately after sending the message, the new function blocks until
  receiving an Ack from the other party on the websocket. It takes a
  context so that the block can be cancelled and force the function to
  return early.

We need this function for usage in the FUD plugin because we don't want
the plugin to terminate before the bastion has acknowledged that it has
received all download data on its end.

Furthermore, by waiting for an Ack, we gain the ability to manage flow
control of messages sent in the websocket. Given that we are going to
increase the write buffer message size and send a lot of messages during
Fud-Download, the websocket can easily get overloaded with a burst of
large messages before the receiving party can process them (i.e.
producer is faster than consumer). Essentially, adding the ability to
wait for an Ack allows us to rate limit ourselves in how fast we send
file data.

It is also important to keep in mind the default limits for the
OutgoingMessageBufferCapacity as defined here:
https://github.com/bastionzero/bzero-ssm-agent/blob/3d9a6a79605d42f277e9552bb6a7a19efac8d247/agent/session/config/config.go#L45-L49.
When we increase the websocket write buffer message size (to give FUD
better download performance), outgoing stream messages held in the
OutgoingMessageBuffer can be larger than the standard 1KB size used by
the other session plugins. Consumers of
SendStreamDataMessageAndWaitForAck (i.e FUD plugin) will have to ensure
that their rate of sending does not create a memory blow-up of the
OutgoingMessageBuffer. From the referenced code above, "When changing
StreamDataPayloadSize, make corresponding change to buffer capacity to
ensure no more than 100MB of instance memory is used." The easiest way
to ensure no memory blow-up, would be to wait for an Ack on _every_ sent
message via SendStreamDataMessageAndWaitForAck before sending another
message at the cost of performance. Performance can be improved by only
waiting for an Ack on every x messages, where x is a number > 1. Again,
it is up to the consumer of this function to figure out a number for x
that isn't so large that it causes memory blow-up.