Add ability to activate FUD plugin
Unfortunately, it is not possible to activate the FUD plugin in the
standard way. When using the port plugin or the interactive command
plugin, activation is done by issuing a special session document to the
AWS SSM API. AWS allows us to create custom documents, but it does not
allow us to change the JSON structure of such documents. Therefore, we
cannot follow the implementation of built-in plugins and create a new
document whose "sessionType" parameter is equal to "FileUploadDownload".

To get around this limitation, we activate FUD by coopting the
InteractiveShellCommand document's "Commands" parameter. This field can
contain any string we choose. For FUD activation, we encode a special
FUD JSON string recognized by the agent.