Change the way the plugin terminates when upload/download is finished
Previously, once the download or upload goroutine finished successfully,
the plugin would terminate immediately. This disconnect event is
actually subscribed to by the bastion and relayed back to the client.

With the addition of web workers in the webapp, a race condition was
introduced (or perhaps it was always there and is worsened) where the
plugin sends all download packets, disconnects, but the client isn't
done processing. The client interpreted the plugin disconnect as an
error even though it still had processing todo.

To fix this and to simplify plugin execution, the plugin now waits for
the context to be cancelled before terminating. Essentially, plugin
termination is now driven by the client/bastion with this change, thus
removing the race condition. The only other way for the plugin to be
terminated is for the session-worker to be killed which is caught by the
sig channel in execute() and also cancels the context.