Add more explicit error handling in fud/upload
Previously, an error in creating the tmp file that checks if the target
user has permission was interpreted always as a permission error. This
commit adds explicit checks for the permission error.

In addition, it checks for other errors such as the path containing
folders that do not exist (parity with scp), or if the destination path
itself is a folder.

+ Adds new FUD error: FUDInvalidDestinationPath