Agent Self-Registration Revamp (#59)
* Pass in Registration Data

Registration flag now allows us to specify a bzero registration route which takes in a json argument we specify and is then checked by membership to a struct.  We store it in the vault, do the rest of the bzero stuff and move on.

* Agent Self-Registers

- Agent goes through the old process of registration with information from reg api call response
- moved orgID and orgProvider from registration flags to being set after we get the registration information response since it's included already.
- All startup messages now say bzero-ssm-agent instead of amazon-ssm-agent
- Added special error code for Bzero Registration Errors

* Modify service files

- Attempt restarts ONLY when there is an unclean error (like a panic or something).  This way registration can execute and exit cleanly without constant restarts.
- Only attempt to restart 3 times, we don't want our service continually running if it's unable to register or there's some flaw.

* Added Robustness to Agent Registration

- Add retry parameter to optionally skip retries
- Add a lot of specific error messages surrounding the API call
- If a user ever tries to register and provides valid registration information, it will clear the old aws registration so that it can redo it.

* Change AWS warn message to debug

I don't think this is ever necessary and happens every time you try to check registration before the agent has ever registered, but I didn't want to get rid of it so I set it to debug just in case it causes issues in the future.

- Changing really minor error reporting
- Fix bug where I accidentally removed a key that let the client understand the bzeroinfo output

* Maintain old registration path and add exponential backoff

- BZero inits on register even without bzero info
- Added exponential backoff to register API call

* Allow trace logs turned on

* Bump version

* Change systemd file

Change restart back to always, just in case although will not restart on a bzero reg error or reboot requested.  Restarts are still limited to 3.

* increase registration retries

* Add new registration request variable for clarity

* Revamp

Code to comply with new requirements for agent self-registration

* Add EnvId and ServiceUrl

Make sure we only assign the OrgID and OrgProvider if they were not provided by the user as opposed to overwriting any existing value.

* Add logging to registration lib

* Add omitempty to json options in registration

If user doesn't specify EnvName or EnvID, omit from registration api request

* Restart Agent after successful registration

Once an agent registers successfully, the agent will restart

* rename apiKey flag to registrationKey

* v2 api change